back-end/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
python-multipart

